# openapi.yaml
openapi: 3.1.0
info:
  title: Medical Supply Web Project API
  version: 1.0.0
  description: |
    Drug licensing, medicine unit serialization, custody transfer, and public
    authenticity verification.

    ## Authentication
    All endpoints except `GET /verify/{unitId}` require a Supabase Auth session cookie.
    The caller's organization and role are resolved server-side from their membership —
    they are never taken from the request. A client cannot name the organization it is
    acting for.

    ## A note on 404
    A request for a resource belonging to another organization returns **404, not 403**.
    A 403 would confirm the resource exists, which is itself a disclosure. Refusals
    about other organizations' data are indistinguishable from absence.

    ## What this API never returns
    No endpoint discloses the identity of a person who purchased medicine. Dispensing
    records the pharmacy and the timestamp; the buyer is not modelled, not stored, and
    not retrievable.
servers:
  - url: /api/v1
...
paths:
...
  /batches/{batchId}/units:
    get:
      tags: [Manufacturer]
      summary: The unit ids of a batch (paged) — this is what the QR export reads
      parameters:
        - { name: batchId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 500, maximum: 1000 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/Refused" }
...
  /verify/{unitId}:
    get:
      tags: [Public]
      summary: Verify a medicine by its code (VER-1/2/3) — no authentication
      description: |
        The only unauthenticated endpoint. Rate-limited to 30 requests per minute per
        caller, and CDN-cached for 60s.

        Every scan is recorded — including scans of codes that were never issued, which
        is the strongest counterfeit signal available. Callers are identified only by a
        salted hash of their IP; no personal data is stored.

        A malformed code returns `200 {"verdict":"not_found"}`, not `400`. A frightened
        person holding a suspicious box should get a clear answer, not an error — and a
        distinct error would let a caller tell "bad shape" from "not one of ours".
      parameters:
        - { name: unitId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicVerification" }
        "429":
          description: Too many requests
          headers:
            Retry-After: { schema: { type: integer } }
components:
...
  responses:
    Refused:
      description: |
        Not found — or found, but not yours. These are deliberately indistinguishable.
  schemas:
...
    PublicVerification:
      type: object
      description: |
        Note what is absent: there is no owner, no buyer, no person of any kind. The
        custody chain names organizations only.
      properties:
        verdict:
          type: string
          enum: [authentic, not_found, flagged]
          description: |
            `flagged` means the unit is real but its scan pattern is anomalous — the
            signature of a code photocopied onto counterfeit packaging.
        unitId: { type: string, format: uuid }
        drug:
          type: object
          properties:
            name: { type: string }
            code: { type: string }
        manufacturer:
          type: object
          properties:
            name: { type: string }
        expirationDate: { type: string, format: date }
        status:
          type: string
          enum: [active, in_transit, dispensed, expired, recalled]
        chain:
          type: array
          description: Custody by organization. Produced -> received -> dispensed.
          items:
            type: object
            properties:
              org: { type: string }
              at: { type: string, format: date-time }
              event: { type: string, enum: [produced, received, dispensed] }
        dispensedBy:
          type: object
          description: The pharmacy and the moment. Never the person.
          properties:
            name: { type: string }
            at: { type: string, format: date-time }
        warnings:
          type: array
          items: { type: string }
...